requirements.txt
fastapi
uvicorn
pytesseract
python-multipart
Pillow
